refactor(backend): split numpy_backend.py into category mixin modules (Phase 3.1)
numpy_backend.py (1256 lines) mixed identity, creation, indexing, math,
linalg, interpolation, random, and misc operations in one file. Split
into same-directory _numpy_*.py mixin modules by category; NumpyBackend
now composes them via multiple inheritance with identical MRO/behavior.
Pure file reorganization -- no method moved logic, only location. Full
test suite (7258 tests) and golden-value regression snapshots pass
unchanged at both backends.
